RAG-Assistant/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-core
langchain-text-splitters
langchain-classic
faiss-cpu
PyPDF2
